requirements.txt
flask==2.3.3
gunicorn==21.2.0
google-cloud-pubsub==2.18.4
neo4j==5.12.0
structlog==23.1.0
python-dotenv==1.0.0
google-auth==2.25.2
google-cloud-logging==3.8.0 